requirements.txt
# Simple Raspberry Pi Fart Detector Dependencies

# Pin NumPy to 1.x (TF wheels aren't ready for NumPy 2.x yet)
numpy==1.26.4

# Audio input
sounddevice==0.5.2

# Audio processing
librosa==0.10.1
scipy==1.11.4

# TensorFlow (aarch64 build, no libclang)
tensorflow-aarch64==2.16.1

# TensorFlow Hub (fallback for YAMNet) - optimized for RPi4 ARM64
tensorflow-hub==0.12.0

# Core runtime deps (safe, pinned for compatibility)
absl-py==2.3.1
protobuf==3.20.3
opt-einsum==3.4.0
typing-extensions==4.15.0
wrapt==1.17.3
six==1.17.0
gast==0.6.0

